refactor for file upload response handlers API
